requirements.txt

Flask-SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
opencv-python
opencv-python-headless
numpy
matplotlib
pillow
gaze-tracking
fer
reportlab
